refactor to link folders

+++ install.sh
@@ -9,10 +9,8 @@ source_files=(
     "$SOURCE_PATH/yofi/yofi.config"
     "$SOURCE_PATH/tmux/.tmux.conf"
     "$SOURCE_PATH/zsh/.zshrc"
-    "$SOURCE_PATH/tofi/config"
-    "$SOURCE_PATH/waybar/config"
-    "$SOURCE_PATH/waybar/style.css"
-
+    "$SOURCE_PATH/tofi"
+    "$SOURCE_PATH/waybar"
 )
 
 dest_files=(
@@ -21,9 +19,8 @@ dest_files=(
     "$CONFIG_PATH/yofi/yofi.config"
     "$HOME/.tmux.conf"
     "$HOME/.zshrc"
-    "$CONFIG_PATH/tofi/config"
-    "$CONFIG_PATH/waybar/config"
-    "$CONFIG_PATH/waybar/style.css"
+    "$CONFIG_PATH/"
+    "$CONFIG_PATH/"
 )
 
 BACKUPS_CREATED=()
@@ -35,8 +32,13 @@ for ((i=0;i<${#source_files[@]};i++)) do
     if [  -L ${dest_files[i]} ]; then
         echo "Is already a link, no action: ${dest_files[i]}"
     else
-        mkdir -p $(dirname ${dest_files[i]})
-        ln -bs ${source_files[i]} ${dest_files[i]}
+        if [ -f ${source_files[i]} ]; then
+            mkdir -p $(dirname ${dest_files[i]})
+            ln -bs ${source_files[i]} ${dest_files[i]}
+        elif [ -d ${source_files[i]} ]; then
+            # todo if dir exists needs backup
+            ln -s ${source_files[i]} ${dest_files[i]}
+        fi
     fi
 
     BACKUP_PATH=${dest_files[i]}~
